+++ RandomSkunk.Results/TryCatch.cs
@@ -31,6 +31,10 @@ public static class TryCatch
             sourceDelegate();
             return Result.Success();
         }
+        catch (TaskCanceledException ex)
+        {
+            return Errors.Canceled(ex);
+        }
         catch (Exception ex)
         {
             return Result.Fail(exceptionHandler(ex));
@@ -59,6 +63,10 @@ public static class TryCatch
         {
             await sourceDelegate().ConfigureAwait(ContinueOnCapturedContext);
             return Result.Success();
+        }
+        catch (TaskCanceledException ex)
+        {
+            return Errors.Canceled(ex);
         }
         catch (Exception ex)
         {
@@ -92,6 +100,10 @@ public static class TryCatch
             var value = sourceDelegate();
             return Result<T>.FromValue(value);
         }
+        catch (TaskCanceledException ex)
+        {
+            return Errors.Canceled(ex);
+        }
         catch (Exception ex)
         {
             return Result<T>.Fail(exceptionHandler(ex));
@@ -123,6 +135,10 @@ public static class TryCatch
         {
             var value = await sourceDelegate().ConfigureAwait(ContinueOnCapturedContext);
             return Result<T>.FromValue(value);
+        }
+        catch (TaskCanceledException ex)
+        {
+            return Errors.Canceled(ex);
         }
         catch (Exception ex)
         {
@@ -156,6 +172,10 @@ public static class TryCatch
             var value = sourceDelegate();
             return Maybe<T>.FromValue(value);
         }
+        catch (TaskCanceledException ex)
+        {
+            return Errors.Canceled(ex);
+        }
         catch (Exception ex)
         {
             return Maybe<T>.Fail(exceptionHandler(ex));
@@ -187,6 +207,10 @@ public static class TryCatch
         {
             var value = await sourceDelegate().ConfigureAwait(ContinueOnCapturedContext);
             return Maybe<T>.FromValue(value);
+        }
+        catch (TaskCanceledException ex)
+        {
+            return Errors.Canceled(ex);
         }
         catch (Exception ex)
         {