+++ RandomSkunk.Results/TryCatch{TException1,TException2,TException3,TException4}.cs
@@ -62,6 +62,10 @@ public static class TryCatch<TException1, TException2, TException3, TException4>
             sourceDelegate();
             return Result.Success();
         }
+        catch (TaskCanceledException ex)
+        {
+            return Errors.Canceled(ex);
+        }
         catch (TException1 ex)
         {
             return Result.Fail(exception1Handler(ex));
@@ -120,6 +124,10 @@ public static class TryCatch<TException1, TException2, TException3, TException4>
         {
             await sourceDelegate().ConfigureAwait(ContinueOnCapturedContext);
             return Result.Success();
+        }
+        catch (TaskCanceledException ex)
+        {
+            return Errors.Canceled(ex);
         }
         catch (TException1 ex)
         {
@@ -183,6 +191,10 @@ public static class TryCatch<TException1, TException2, TException3, TException4>
             var value = sourceDelegate();
             return Result<T>.FromValue(value);
         }
+        catch (TaskCanceledException ex)
+        {
+            return Errors.Canceled(ex);
+        }
         catch (TException1 ex)
         {
             return Result<T>.Fail(exception1Handler(ex));
@@ -244,6 +256,10 @@ public static class TryCatch<TException1, TException2, TException3, TException4>
         {
             var value = await sourceDelegate().ConfigureAwait(ContinueOnCapturedContext);
             return Result<T>.FromValue(value);
+        }
+        catch (TaskCanceledException ex)
+        {
+            return Errors.Canceled(ex);
         }
         catch (TException1 ex)
         {
@@ -307,6 +323,10 @@ public static class TryCatch<TException1, TException2, TException3, TException4>
             var value = sourceDelegate();
             return Maybe<T>.FromValue(value);
         }
+        catch (TaskCanceledException ex)
+        {
+            return Errors.Canceled(ex);
+        }
         catch (TException1 ex)
         {
             return Maybe<T>.Fail(exception1Handler(ex));
@@ -368,6 +388,10 @@ public static class TryCatch<TException1, TException2, TException3, TException4>
         {
             var value = await sourceDelegate().ConfigureAwait(ContinueOnCapturedContext);
             return Maybe<T>.FromValue(value);
+        }
+        catch (TaskCanceledException ex)
+        {
+            return Errors.Canceled(ex);
         }
         catch (TException1 ex)
         {