+++ RandomSkunk.Results/TryCatch{TException1,TException2,TException3}.cs
@@ -53,6 +53,10 @@ public static class TryCatch<TException1, TException2, TException3>
             sourceDelegate();
             return Result.Success();
         }
+        catch (TaskCanceledException ex)
+        {
+            return Errors.Canceled(ex);
+        }
         catch (TException1 ex)
         {
             return Result.Fail(exception1Handler(ex));
@@ -102,6 +106,10 @@ public static class TryCatch<TException1, TException2, TException3>
         {
             await sourceDelegate().ConfigureAwait(ContinueOnCapturedContext);
             return Result.Success();
+        }
+        catch (TaskCanceledException ex)
+        {
+            return Errors.Canceled(ex);
         }
         catch (TException1 ex)
         {
@@ -155,6 +163,10 @@ public static class TryCatch<TException1, TException2, TException3>
             var value = sourceDelegate();
             return Result<T>.FromValue(value);
         }
+        catch (TaskCanceledException ex)
+        {
+            return Errors.Canceled(ex);
+        }
         catch (TException1 ex)
         {
             return Result<T>.Fail(exception1Handler(ex));
@@ -206,6 +218,10 @@ public static class TryCatch<TException1, TException2, TException3>
         {
             var value = await sourceDelegate().ConfigureAwait(ContinueOnCapturedContext);
             return Result<T>.FromValue(value);
+        }
+        catch (TaskCanceledException ex)
+        {
+            return Errors.Canceled(ex);
         }
         catch (TException1 ex)
         {
@@ -259,6 +275,10 @@ public static class TryCatch<TException1, TException2, TException3>
             var value = sourceDelegate();
             return Maybe<T>.FromValue(value);
         }
+        catch (TaskCanceledException ex)
+        {
+            return Errors.Canceled(ex);
+        }
         catch (TException1 ex)
         {
             return Maybe<T>.Fail(exception1Handler(ex));
@@ -310,6 +330,10 @@ public static class TryCatch<TException1, TException2, TException3>
         {
             var value = await sourceDelegate().ConfigureAwait(ContinueOnCapturedContext);
             return Maybe<T>.FromValue(value);
+        }
+        catch (TaskCanceledException ex)
+        {
+            return Errors.Canceled(ex);
         }
         catch (TException1 ex)
         {