+++ RandomSkunk.Results/TryCatch{TException1,TException2}.cs
@@ -44,6 +44,10 @@ public static class TryCatch<TException1, TException2>
             sourceDelegate();
             return Result.Success();
         }
+        catch (TaskCanceledException ex)
+        {
+            return Errors.Canceled(ex);
+        }
         catch (TException1 ex)
         {
             return Result.Fail(exception1Handler(ex));
@@ -84,6 +88,10 @@ public static class TryCatch<TException1, TException2>
         {
             await sourceDelegate().ConfigureAwait(ContinueOnCapturedContext);
             return Result.Success();
+        }
+        catch (TaskCanceledException ex)
+        {
+            return Errors.Canceled(ex);
         }
         catch (TException1 ex)
         {
@@ -128,6 +136,10 @@ public static class TryCatch<TException1, TException2>
             var value = sourceDelegate();
             return Result<T>.FromValue(value);
         }
+        catch (TaskCanceledException ex)
+        {
+            return Errors.Canceled(ex);
+        }
         catch (TException1 ex)
         {
             return Result<T>.Fail(exception1Handler(ex));
@@ -170,6 +182,10 @@ public static class TryCatch<TException1, TException2>
         {
             var value = await sourceDelegate().ConfigureAwait(ContinueOnCapturedContext);
             return Result<T>.FromValue(value);
+        }
+        catch (TaskCanceledException ex)
+        {
+            return Errors.Canceled(ex);
         }
         catch (TException1 ex)
         {
@@ -214,6 +230,10 @@ public static class TryCatch<TException1, TException2>
             var value = sourceDelegate();
             return Maybe<T>.FromValue(value);
         }
+        catch (TaskCanceledException ex)
+        {
+            return Errors.Canceled(ex);
+        }
         catch (TException1 ex)
         {
             return Maybe<T>.Fail(exception1Handler(ex));
@@ -256,6 +276,10 @@ public static class TryCatch<TException1, TException2>
         {
             var value = await sourceDelegate().ConfigureAwait(ContinueOnCapturedContext);
             return Maybe<T>.FromValue(value);
+        }
+        catch (TaskCanceledException ex)
+        {
+            return Errors.Canceled(ex);
         }
         catch (TException1 ex)
         {